+++ test/function/construction/string.test.js
@@ -31,8 +31,8 @@ describe('string', function() {
   it('should convert a number to string', function() {
     assert.equal(string(1/8), '0.125');
     assert.equal(string(2.1e-3), '0.0021');
-    assert.equal(string(123456789), '1.23456789e8');
-    assert.equal(string(2000000), '2e6');
+    assert.equal(string(123456789), '1.23456789e+8');
+    assert.equal(string(2000000), '2e+6');
   });
 
   it('should convert a complex number to string', function() {